ec2_snapshot: enable permissioning
Allow createVolumePermission snapshot attributes to be set.

Update tagging to be idempotent and correctly report changes.

Use Ansible's built in parameter validation methods (`required_if`,
`mutually_exclusive`, `required_together`).

Ensure check_mode works correctly

+++ lib/ansible/modules/cloud/amazon/ec2_snapshot.py
@@ -71,7 +71,7 @@ options:
     version_added: "1.9"
   snapshot_id:
     description:
-      - snapshot id to remove
+      - snapshot id to remove or modify
     required: false
     version_added: "1.9"
   last_snapshot_min_age:
@@ -80,6 +80,14 @@ options:
     required: false
     default: 0
     version_added: "2.0"
+  create_volume_permissions:
+    description:
+      - Users and groups that should be able to create a volume from a snapshot.
+        Dictionary with a key of user_ids and/or group_names. user_ids should
+        be a list of account ids. group_name should be a list of groups, "all"
+        is the only acceptable value currently.
+    required: false
+    version_added: "2.4"
 
 author: "Will Thames (@willthames)"
 extends_documentation_fragment:
@@ -108,20 +116,28 @@ EXAMPLES = '''
         source: /data
 
 # Remove a snapshot
-- local_action:
-    module: ec2_snapshot
+- ec2_snapshot:
     snapshot_id: snap-abcd1234
     state: absent
 
 # Create a snapshot only if the most recent one is older than 1 hour
-- local_action:
-    module: ec2_snapshot
+- ec2_snapshot:
     volume_id: vol-abcdef12
     last_snapshot_min_age: 60
+
+# Allow account 1234567890 to access a snapshot
+- ec2_snapshot:
+    snapshot_id: snap-abcd1234
+    create_volume_permissions:
+      user_ids:
+      - '1234567890'
 '''
 
 import time
 import datetime
+
+from ansible.module_utils.basic import AnsibleModule
+from ansible.module_utils.ec2 import ec2_connect, ec2_argument_spec
 
 try:
     import boto.ec2
@@ -182,18 +198,23 @@ def _create_with_wait(snapshot, wait_timeout_secs, sleep_func=time.sleep):
     return True
 
 
+def existing_to_desired(existing_list, desired_list):
+    """
+       Takes two lists and returns the elements that need
+       to be added to the existing_list and the elements
+       that need to be removed from the existing_list to
+       become the desired_list
+    """
+
+    return list(set(desired_list) - set(existing_list)), list(set(existing_list) - set(desired_list))
+
+
 def create_snapshot(module, ec2, state=None, description=None, wait=None,
                     wait_timeout=None, volume_id=None, instance_id=None,
                     snapshot_id=None, device_name=None, snapshot_tags=None,
-                    last_snapshot_min_age=None):
+                    last_snapshot_min_age=None, create_volume_permissions=None):
     snapshot = None
     changed = False
-
-    required = [volume_id, snapshot_id, instance_id]
-    if required.count(None) != len(required) - 1: # only 1 must be set
-        module.fail_json(msg='One and only one of volume_id or instance_id or snapshot_id must be specified')
-    if instance_id and not device_name or device_name and not instance_id:
-        module.fail_json(msg='Instance ID and device name must both be specified')
 
     if instance_id:
         try:
@@ -207,10 +228,9 @@ def create_snapshot(module, ec2, state=None, description=None, wait=None,
         volume_id = volumes[0].id
 
     if state == 'absent':
-        if not snapshot_id:
-            module.fail_json(msg = 'snapshot_id must be set when state is absent')
         try:
-            ec2.delete_snapshot(snapshot_id)
+            if not module.check_mode:
+                ec2.delete_snapshot(snapshot_id)
         except boto.exception.BotoServerError as e:
             # exception is raised if snapshot does not exist
             if e.error_code == 'InvalidSnapshot.NotFound':
@@ -231,16 +251,57 @@ def create_snapshot(module, ec2, state=None, description=None, wait=None,
         snapshot = _get_most_recent_snapshot(current_snapshots,
                                              max_snapshot_age_secs=last_snapshot_min_age)
     try:
+        if snapshot_id:
+            snapshot = ec2.get_all_snapshots(snapshot_ids=[snapshot_id])[0]
         # Create a new snapshot if we didn't find an existing one to use
         if snapshot is None:
-            snapshot = ec2.create_snapshot(volume_id, description=description)
-            changed = True
+            if not module.check_mode:
+                snapshot = ec2.create_snapshot(volume_id, description=description)
+                changed = True
+            else:
+                module.exit_json(changed=True)
         if wait:
             if not _create_with_wait(snapshot, wait_timeout):
                 module.fail_json(msg='Timed out while creating snapshot.')
-        if snapshot_tags:
-            for k, v in snapshot_tags.items():
+
+        tags_to_add, tags_to_remove = existing_to_desired(snapshot.tags.items(), snapshot_tags.items())
+        for (k, v) in tags_to_add:
+            if not module.check_mode:
                 snapshot.add_tag(k, v)
+            changed = True
+        for (k, v) in tags_to_remove:
+            if not module.check_mode:
+                snapshot.remove_tag(k, value=v)
+            changed = True
+
+        if module.check_mode:
+            tags = snapshot_tags.copy()
+        else:
+            tags = snapshot.tags.copy()
+
+        permissions = ec2.get_snapshot_attribute(snapshot.id)
+
+        users_to_add, users_to_remove = existing_to_desired(permissions.attrs.get('user_ids', []),
+                                                            [str(user_id) for user_id in
+                                                             create_volume_permissions.get('user_ids', [])])
+        groups_to_add, groups_to_remove = existing_to_desired(permissions.attrs.get('groups', []),
+                                                              create_volume_permissions.get('groups', []))
+
+        if users_to_add or groups_to_add:
+            if not module.check_mode:
+                ec2.modify_snapshot_attribute(snapshot.id, user_ids=users_to_add, groups=groups_to_add)
+                permissions = ec2.get_snapshot_attribute(snapshot.id)
+            else:
+                permissions = create_volume_permissions
+            changed = True
+        if users_to_remove or groups_to_remove:
+            if not module.check_mode:
+                ec2.modify_snapshot_attribute(snapshot.id, operation='remove',
+                                              user_ids=users_to_remove, groups=groups_to_remove)
+                permissions = ec2.get_snapshot_attribute(snapshot.id)
+            else:
+                permissions = create_volume_permissions
+            changed = True
     except boto.exception.BotoServerError as e:
         module.fail_json(msg="%s: %s" % (e.error_code, e.error_message))
 
@@ -248,7 +309,8 @@ def create_snapshot(module, ec2, state=None, description=None, wait=None,
                      snapshot_id=snapshot.id,
                      volume_id=snapshot.volume_id,
                      volume_size=snapshot.volume_size,
-                     tags=snapshot.tags.copy())
+                     tags=tags,
+                     permissions=permissions.attrs)
 
 
 def create_snapshot_ansible_module():
@@ -265,9 +327,18 @@ def create_snapshot_ansible_module():
             last_snapshot_min_age = dict(type='int', default=0),
             snapshot_tags = dict(type='dict', default=dict()),
             state = dict(choices=['absent','present'], default='present'),
+            create_volume_permissions=dict(type='dict', default=dict()),
         )
     )
-    module = AnsibleModule(argument_spec=argument_spec)
+    module = AnsibleModule(argument_spec=argument_spec,
+                           supports_check_mode=True,
+                           required_together=[['instance_id', 'device_name']],
+                           required_if=[['state', 'absent', ['snapshot_id']]],
+                           mutually_exclusive=[['volume_id', 'instance_id'],
+                                               ['volume_id', 'snapshot_id'],
+                                               ['instance_id', 'snapshot_id'],
+                                               ['snapshot_id', 'last_snapshot_min_age']],
+                          )
     return module
 
 
@@ -287,6 +358,7 @@ def main():
     last_snapshot_min_age = module.params.get('last_snapshot_min_age')
     snapshot_tags = module.params.get('snapshot_tags')
     state = module.params.get('state')
+    create_volume_permissions = module.params.get('create_volume_permissions')
 
     ec2 = ec2_connect(module)
 
@@ -302,12 +374,10 @@ def main():
         snapshot_id=snapshot_id,
         device_name=device_name,
         snapshot_tags=snapshot_tags,
-        last_snapshot_min_age=last_snapshot_min_age
+        last_snapshot_min_age=last_snapshot_min_age,
+        create_volume_permissions=create_volume_permissions
     )
 
-# import module snippets
-from ansible.module_utils.basic import *
-from ansible.module_utils.ec2 import *
 
 if __name__ == '__main__':
     main()
